+++ R/load_all.R
@@ -24,6 +24,7 @@
 #' If there are many datasources, then it is advised to set this TRUE, as then each different datasource will be processed in parallel.
 #' However, if there are only a few datasources selected to load, but many files per datasource (result of large queries), then it may be faster to parallelize within each datasource and therefore should be set to \emph{FALSE}.
 #' If there are only a few sources each with one file then set to TRUE.
+#' @param load_report boolean, should the report text be returned for notes.
 #'
 #' @return list of parsed data tables containing the information.
 #'
@@ -38,7 +39,7 @@
 #' }
 
 load_all <- function(folder, which_data = c("mrn", "con", "dem", "enc", "rdt", "lab", "med", "dia", "rfv", "prc", "lno", "car", "dis", "end", "hnp", "opn", "pat", "prg", "pul", "rad", "vis"), old_dem = FALSE,
-                     merge_id = "EMPI", sep = ":", id_length = "standard", perc = 0.6, na = TRUE, identical = TRUE, nThread = 4, many_sources = TRUE) {
+                     merge_id = "EMPI", sep = ":", id_length = "standard", perc = 0.6, na = TRUE, identical = TRUE, nThread = 4, many_sources = TRUE, load_report = TRUE) {
 
   .SD=.N=.I=.GRP=.BY=.EACHI=..=..cols=.SDcols=i=j=time_to_db=..which_ids_to=..which_ids_from <- NULL
 
@@ -120,6 +121,7 @@ load_all <- function(folder, which_data = c("mrn", "con", "dem", "enc", "rdt", "
                                                   "id_length = \"", id_length, "\", ",
                                                   "perc = ", perc, ", ",
                                                   "nThread = ", 1, ", ",
+                                                  "load_report = \"", load_report, "\", ",
                                                   "na = FALSE, identical = FALSE)")))
               } else {
                 func <- grep(type, x = tolower(load_functions), value = TRUE, fixed = TRUE)
@@ -215,6 +217,7 @@ load_all <- function(folder, which_data = c("mrn", "con", "dem", "enc", "rdt", "
                                                   "id_length = \"", id_length, "\", ",
                                                   "perc = ", perc, ", ",
                                                   "nThread = ", 1, ", ",
+                                                  "load_report = \"", load_report, "\", ",
                                                   "na = FALSE, identical = FALSE)")))
               } else {
                 func <- grep(type, x = tolower(load_functions), value = TRUE, fixed = TRUE)